+++ libs/langgraph-api/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @langchain/langgraph-api
 
+## 0.0.71
+
+### Patch Changes
+
+- f5865ac: Update API spec to match python api
+  - @langchain/langgraph-ui@0.0.71
+
+## 0.0.70
+
+### Patch Changes
+
+- 636e142: Updated types to match python api
+  - @langchain/langgraph-ui@0.0.70
+
+## 0.0.69
+
+### Patch Changes
+
+- f2aa533: Updated typing of auth filters
+  - @langchain/langgraph-ui@0.0.69
+
 ## 0.0.68
 
 ### Patch Changes