+++ libs/langgraph-cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @langchain/langgraph-cli
 
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies [f5865ac]
+  - @langchain/langgraph-api@0.0.71
+
+## 0.0.70
+
+### Patch Changes
+
+- Updated dependencies [636e142]
+  - @langchain/langgraph-api@0.0.70
+
+## 0.0.69
+
+### Patch Changes
+
+- Updated dependencies [f2aa533]
+  - @langchain/langgraph-api@0.0.69
+
 ## 0.0.68
 
 ### Patch Changes